+++ CHANGELOG.md
@@ -8,7 +8,8 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - client functions `NewDefaultHTTPClient` and `NewDefaultHTTPTransport` to provide HTTP client default properties
 - kubernetes: experimental support for deleting nodes from node groups
-- kubernetes: consts for `scaling-up` and `scaling-down` node-group states
+- kubernetes: consts for `scaling-up` and `scaling-down` node group states
+- kubernetes: `utility_network_access` field to node group for configuring utility network access on the given group
 
 ### Changed
 - `service.GetKubernetesNodeGroup` method to return `upcloud.KubernetesNodeGroupDetails` type which is extended version of the previous `upcloud.KubernetesNodeGroup`